pyproject.toml
[project]
name = "concrete-crack-cnn-vs-vit"
version = "0.1.0"
description = "CNN vs Vision Transformer benchmark for concrete crack detection and segmentation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Tobias Hoffmann", email = "thoffmann-ml@proton.me" }]
dependencies = [
    # Core deep learning
    "torch>=2.2",
    "torchvision>=0.17",
    "lightning>=2.2",
    # Model zoos
    "timm>=1.0",
    "segmentation-models-pytorch>=0.3",
    # Data & augmentation
    "albumentations>=1.4",
    "opencv-python>=4.9",
    "Pillow>=10.0",
    "scikit-image>=0.22",
    # Metrics & tensor ops
    "torchmetrics>=1.3",
    "einops>=0.7",
    # Experiment tracking & config
    "mlflow>=2.10",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    # Export & inference
    "onnx>=1.15",
    "onnxruntime>=1.17",
    # Data science
    "numpy>=1.26",
    "pandas>=2.1",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "scikit-learn>=1.4",
    # DX / logging
    "loguru>=0.7",
    "rich>=13.0",
    "tqdm>=4.66",
    "jupyterlab>=4.5.4",
    "kaggle>=1.7.4.5",
]

[project.urls]
Homepage = "https://github.com/thfmn/concrete-crack-cnn-vs-vit"
Repository = "https://github.com/thfmn/concrete-crack-cnn-vs-vit"
Issues = "https://github.com/thfmn/concrete-crack-cnn-vs-vit/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: end-to-end tests using real images from tests/fixtures/",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
